+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,11 @@
 # API Elements: CLI Changelog
 
+## 0.10.2 (2020-08-05)
+
+This update incorporates changes from API Element Adapters:
+
+- openapi3-parser 0.15.0
+
 ## 0.10.1 (2020-06-24)
 
 This update incorporates changes from API Element Adapters: