+++ packages/openapi2-parser/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - added a Link element to the specific format/version in the parse result.
 
+## 0.32.3 (2020-08-06)
+
+Adds compatibility for @apielements/core 0.2.0.
+
 ## 0.32.2 (2020-07-22)
 
 ### Bug Fixes